move make_flat_df and tests
