+++ tests/pytest_pysd/pytest_utils.py
@@ -99,197 +99,6 @@ class TestUtils():
                 ["inflow_a", "inflow_b", "inflow_c"],
                 {'Inflow A': 'inflow_a', 'Inflow B': 'inflow_b'})
 
-    def test_make_flat_df(self):
-
-        df = pd.DataFrame(index=[1], columns=['elem1'])
-        df.at[1] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                 {'Dim1': ['A', 'B', 'C'],
-                                  'Dim2': ['D', 'E', 'F']},
-                                 dims=['Dim1', 'Dim2'])]
-
-        expected = pd.DataFrame(index=[1], data={'Elem1[B,F]': 6.})
-
-        return_addresses = {
-            'Elem1[B,F]': ('elem1', {'Dim1': ['B'], 'Dim2': ['F']})}
-
-        actual = pysd.utils.make_flat_df(df, return_addresses)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        assert_frames_close(actual, expected, rtol=1e-8, atol=1e-8)
-
-    def test_make_flat_df_0dxarray(self):
-
-        df = pd.DataFrame(index=[1], columns=['elem1'])
-        df.at[1] = [xr.DataArray(5)]
-
-        expected = pd.DataFrame(index=[1], data={'Elem1': 5.})
-
-        return_addresses = {'Elem1': ('elem1', {})}
-
-        actual = pysd.utils.make_flat_df(df, return_addresses, flatten=True)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        assert_frames_close(actual, expected, rtol=1e-8, atol=1e-8)
-
-    def test_make_flat_df_nosubs(self):
-
-        df = pd.DataFrame(index=[1], columns=['elem1', 'elem2'])
-        df.at[1] = [25, 13]
-
-        expected = pd.DataFrame(index=[1], columns=['Elem1', 'Elem2'])
-        expected.at[1] = [25, 13]
-
-        return_addresses = {'Elem1': ('elem1', {}),
-                            'Elem2': ('elem2', {})}
-
-        actual = pysd.utils.make_flat_df(df, return_addresses)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        assert all(actual['Elem1'] == expected['Elem1'])
-        assert all(actual['Elem2'] == expected['Elem2'])
-
-    def test_make_flat_df_return_array(self):
-        """ There could be cases where we want to
-        return a whole section of an array - ie, by passing in only part of
-        the simulation dictionary. in this case, we can't force to float..."""
-
-        df = pd.DataFrame(index=[1], columns=['elem1', 'elem2'])
-        df.at[1] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                 {'Dim1': ['A', 'B', 'C'],
-                                  'Dim2': ['D', 'E', 'F']},
-                                 dims=['Dim1', 'Dim2']),
-                    xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                 {'Dim1': ['A', 'B', 'C'],
-                                  'Dim2': ['D', 'E', 'F']},
-                                 dims=['Dim1', 'Dim2'])]
-
-        expected = pd.DataFrame(index=[1], columns=['Elem1[A, Dim2]', 'Elem2'])
-        expected.at[1] = [xr.DataArray([[1, 2, 3]],
-                                       {'Dim1': ['A'],
-                                        'Dim2': ['D', 'E', 'F']},
-                                       dims=['Dim1', 'Dim2']),
-                          xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                       {'Dim1': ['A', 'B', 'C'],
-                                        'Dim2': ['D', 'E', 'F']},
-                                       dims=['Dim1', 'Dim2'])]
-
-        return_addresses = {
-            'Elem1[A, Dim2]': ('elem1', {'Dim1': ['A'],
-                                         'Dim2': ['D', 'E', 'F']}),
-            'Elem2': ('elem2', {})}
-
-        actual = pysd.utils.make_flat_df(df, return_addresses)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        # need to assert one by one as they are xarrays
-        assert actual.loc[1, 'Elem1[A, Dim2]'].equals(
-                expected.loc[1, 'Elem1[A, Dim2]'])
-        assert actual.loc[1, 'Elem2'].equals(expected.loc[1, 'Elem2'])
-
-    def test_make_flat_df_flatten(self):
-
-        df = pd.DataFrame(index=[1], columns=['elem1', 'elem2'])
-        df.at[1] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                 {'Dim1': ['A', 'B', 'C'],
-                                  'Dim2': ['D', 'E', 'F']},
-                                 dims=['Dim1', 'Dim2']),
-                    xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                 {'Dim1': ['A', 'B', 'C'],
-                                  'Dim2': ['D', 'E', 'F']},
-                                 dims=['Dim1', 'Dim2'])]
-
-        expected = pd.DataFrame(index=[1], columns=[
-            'Elem1[A,D]',
-            'Elem1[A,E]',
-            'Elem1[A,F]',
-            'Elem2[A,D]',
-            'Elem2[A,E]',
-            'Elem2[A,F]',
-            'Elem2[B,D]',
-            'Elem2[B,E]',
-            'Elem2[B,F]',
-            'Elem2[C,D]',
-            'Elem2[C,E]',
-            'Elem2[C,F]'])
-
-        expected.at[1] = [1, 2, 3, 1, 2, 3, 4, 5, 6, 7, 8, 9]
-
-        return_addresses = {
-            'Elem1[A,Dim2]': ('elem1', {'Dim1': ['A'],
-                                        'Dim2': ['D', 'E', 'F']}),
-            'Elem2': ('elem2', {})}
-
-        actual = pysd.utils.make_flat_df(df, return_addresses, flatten=True)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        # need to assert one by one as they are xarrays
-        for col in set(expected.columns):
-            assert actual.loc[:, col].values == expected.loc[:, col].values
-
-    def test_make_flat_df_flatten_transposed(self):
-
-        df = pd.DataFrame(index=[1], columns=['elem2'])
-        df.at[1] = [
-            xr.DataArray(
-                [[1, 4, 7], [2, 5, 8], [3, 6, 9]],
-                {'Dim2': ['D', 'E', 'F'], 'Dim1': ['A', 'B', 'C']},
-                ['Dim2', 'Dim1']
-            ).transpose("Dim1", "Dim2")
-        ]
-
-        expected = pd.DataFrame(index=[1], columns=[
-            'Elem2[A,D]',
-            'Elem2[A,E]',
-            'Elem2[A,F]',
-            'Elem2[B,D]',
-            'Elem2[B,E]',
-            'Elem2[B,F]',
-            'Elem2[C,D]',
-            'Elem2[C,E]',
-            'Elem2[C,F]'])
-
-        expected.at[1] = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-
-        return_addresses = {
-            'Elem2': ('elem2', {})}
-
-        actual = pysd.utils.make_flat_df(df, return_addresses, flatten=True)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        # need to assert one by one as they are xarrays
-        for col in set(expected.columns):
-            assert actual.loc[:, col].values == expected.loc[:, col].values
-
-    def test_make_flat_df_times(self):
-
-        df = pd.DataFrame(index=[1, 2], columns=['elem1'])
-        df['elem1'] = [xr.DataArray([[1, 2, 3], [4, 5, 6], [7, 8, 9]],
-                                    {'Dim1': ['A', 'B', 'C'],
-                                    'Dim2': ['D', 'E', 'F']},
-                                    dims=['Dim1', 'Dim2']),
-                       xr.DataArray([[2, 4, 6], [8, 10, 12], [14, 16, 19]],
-                                    {'Dim1': ['A', 'B', 'C'],
-                                     'Dim2': ['D', 'E', 'F']},
-                                    dims=['Dim1', 'Dim2'])]
-
-        expected = pd.DataFrame([{'Elem1[B,F]': 6}, {'Elem1[B,F]': 12}])
-        expected.index = [1, 2]
-
-        return_addresses = {'Elem1[B,F]': ('elem1', {'Dim1': ['B'],
-                                                     'Dim2': ['F']})}
-        actual = pysd.utils.make_flat_df(df, return_addresses)
-
-        # check all columns are in the DataFrame
-        assert set(actual.columns) == set(expected.columns)
-        assert set(actual.index) == set(expected.index)
-        assert all(actual['Elem1[B,F]'] == expected['Elem1[B,F]'])
-
     def test_doctests(self):
         doctest.DocTestSuite(pysd.utils)
